requirements.txt
openai
python-dotenv
pypdf
markdownify
python-docx